+++ src/libsac2c/tree/tree_compound.h
@@ -854,7 +854,7 @@ extern node *TCnodeBehindCast (node *arg_node);
 extern node *TCmakeVector (ntype *basetype, node *aelems);
 extern node *TCmakeIntVector (node *aelems);
 extern node *TCcreateIntVector (int length, int value, int step);
-extern int TCgetIntVectorNthValue (int pos, node *vect);
+extern int TCgetIntVectorNthValue (size_t pos, node *vect);
 
 extern node *TCcreateZeroScalar (simpletype btype);
 extern node *TCcreateZeroVector (int length, simpletype btype);
